+++ Shapes/ColorPallet.cpp
@@ -36,6 +36,6 @@ double ColorPallet::getHeight()
 	return 0.0;
 }
 
-void ColorPallet::resize(int factor) {
+void ColorPallet::resize(float factor) {
 
 }
