+++ Shapes/StickImage.cpp
@@ -48,6 +48,6 @@ double StickImage::getHeight()
 {
 	return 0.0;
 }
-void StickImage::resize(int factor) {
+void StickImage::resize(float factor) {
 
 }
