+++ Shapes/Triangle.h
@@ -20,5 +20,5 @@ public:
 	void LOAD(ifstream& Infile);
 	double getWidth();
 	double getHeight();
-	void resize(int factor) override;
+	void resize(float factor) override;
 };
